Convert custom panel to typescript (#2991)
* Convert custom panel to typescript

* Address comments
