+++ src/data/panel_custom.ts
@@ -1,0 +1,8 @@
+export interface CustomPanelConfig {
+  name: string;
+  embed_iframe: boolean;
+  trust_external: boolean;
+  js_url?: string;
+  module_url?: string;
+  html_url?: string;
+}
