+++ src/entrypoints/custom-panel.ts
@@ -1,0 +1,97 @@
+import { loadJS } from "../common/dom/load_resource";
+import { loadCustomPanel } from "../util/custom-panel/load-custom-panel";
+import { createCustomPanelElement } from "../util/custom-panel/create-custom-panel-element";
+import { setCustomPanelProperties } from "../util/custom-panel/set-custom-panel-properties";
+import { fireEvent } from "../common/dom/fire_event";
+import { navigate } from "../common/navigate";
+import { PolymerElement } from "@polymer/polymer";
+import { Panel } from "../types";
+import { CustomPanelConfig } from "../data/panel_custom";
+
+declare global {
+  interface Window {
+    loadES5Adapter: () => Promise<unknown>;
+  }
+}
+
+const webComponentsSupported =
+  "customElements" in window &&
+  "import" in document.createElement("link") &&
+  "content" in document.createElement("template");
+
+let es5Loaded: Promise<unknown> | undefined;
+
+window.loadES5Adapter = () => {
+  if (!es5Loaded) {
+    es5Loaded = Promise.all([
+      loadJS(`${__STATIC_PATH__}custom-elements-es5-adapter.js`).catch(),
+      import(/* webpackChunkName: "compat" */ "./compatibility"),
+    ]);
+  }
+  return es5Loaded;
+};
+
+let panelEl: HTMLElement | PolymerElement | undefined;
+
+function setProperties(properties) {
+  if (!panelEl) {
+    return;
+  }
+  setCustomPanelProperties(panelEl, properties);
+}
+
+function initialize(panel: Panel, properties: {}) {
+  const style = document.createElement("style");
+  style.innerHTML = "body{margin:0}";
+  document.head.appendChild(style);
+
+  const config = panel.config!._panel_custom as CustomPanelConfig;
+  let start: Promise<unknown> = Promise.resolve();
+
+  if (!webComponentsSupported) {
+    start = start.then(() => loadJS("/static/webcomponents-bundle.js"));
+  }
+
+  if (__BUILD__ === "es5") {
+    // Load ES5 adapter. Swallow errors as it raises errors on old browsers.
+    start = start.then(() => window.loadES5Adapter());
+  }
+
+  start
+    .then(() => loadCustomPanel(config))
+    // If our element is using es5, let it finish loading that and define element
+    // This avoids elements getting upgraded after being added to the DOM
+    .then(() => es5Loaded || Promise.resolve())
+    .then(
+      () => {
+        panelEl = createCustomPanelElement(config);
+
+        const forwardEvent = (ev) => {
+          if (window.parent.customPanel) {
+            fireEvent(window.parent.customPanel, ev.type, ev.detail);
+          }
+        };
+        panelEl!.addEventListener("hass-toggle-menu", forwardEvent);
+        window.addEventListener("location-changed", (ev: any) =>
+          navigate(
+            window.parent.customPanel,
+            window.location.pathname,
+            ev.detail ? ev.detail.replace : false
+          )
+        );
+        setProperties({ panel, ...properties });
+        document.body.appendChild(panelEl!);
+      },
+      (err) => {
+        // tslint:disable-next-line
+        console.error(err, panel);
+        alert(`Unable to load the panel source: ${err}.`);
+      }
+    );
+}
+
+document.addEventListener(
+  "DOMContentLoaded",
+  () => window.parent.customPanel!.registerIframe(initialize, setProperties),
+  { once: true }
+);
