+++ src/util/custom-panel/create-custom-panel-element.ts
@@ -1,8 +1,8 @@
-export default function createCustomPanelElement(panelConfig) {
+export const createCustomPanelElement = (panelConfig) => {
   // Legacy support. Custom panels used to have to define element ha-panel-{name}
   const tagName =
     "html_url" in panelConfig
       ? `ha-panel-${panelConfig.name}`
       : panelConfig.name;
   return document.createElement(tagName);
-}
+};
