+++ src/util/custom-panel/load-custom-panel.ts
@@ -3,7 +3,7 @@ import { loadJS, loadModule } from "../../common/dom/load_resource";
 // Make sure we only import every JS-based panel once (HTML import has this built-in)
 const JS_CACHE = {};
 
-export default function loadCustomPanel(panelConfig) {
+export const loadCustomPanel = (panelConfig): Promise<unknown> => {
   if (panelConfig.html_url) {
     const toLoad = [
       import(/* webpackChunkName: "import-href-polyfill" */ "../../resources/html-import/import-href"),
@@ -29,4 +29,4 @@ export default function loadCustomPanel(panelConfig) {
     return loadModule(panelConfig.module_url);
   }
   return Promise.reject("No valid url found in panel config.");
-}
+};
